Implement expand2text
This finishes implementing `expand2text`, and moves all React-specific
code to expand2react.js.
