+++ root/static/scripts/tests/i18n/expand2.js
@@ -4,7 +4,7 @@ import expand2html from '../../common/i18n/expand2html';
 import expand2text from '../../common/i18n/expand2text';
 
 test('expand2', function (t) {
-  t.plan(55);
+  t.plan(63);
 
   let error = '';
   const consoleError = console.error;
@@ -26,7 +26,7 @@ test('expand2', function (t) {
   expandText(null, null, '');
   expandText(undefined, null, '');
   expandText('Some plain text', null, 'Some plain text');
-  // TODO: expandText('Some &quot;plain&quot; text', null, 'Some &quot;plain&quot; text');
+  expandText('Some &quot;plain&quot; text', null, 'Some &quot;plain&quot; text');
   expandHtml('Some &quot;plain&quot; text', null, 'Some &quot;plain&quot; text');
   expandText('An {apple_fruit}', null, 'An {apple_fruit}');
   expandText('An {apple_fruit}', {apple_fruit: 'apple'}, 'An apple');
@@ -42,19 +42,18 @@ test('expand2', function (t) {
     'An <b>apple</b>',
   );
 
-  // TODO: Shouldn't interpolate React elements with expand2text.
-  // expandText(
-  //   'An {apple_fruit}',
-  //   {apple_fruit: React.createElement('b', null, 'apple')},
-  //   'An ',
-  // );
+  // Shouldn't interpolate React elements with expand2text.
+  expandText(
+    'An {apple_fruit}',
+    {apple_fruit: React.createElement('b', null, 'apple')},
+    'An ',
+  );
 
-  // TODO:
-  // expandText(
-  //   'An &lbrace;apple_fruit&rbrace;',
-  //   {apple_fruit: 'apple'},
-  //   'An &lbrace;apple_fruit&rbrace;',
-  // );
+  expandText(
+    'An &lbrace;apple_fruit&rbrace;',
+    {apple_fruit: 'apple'},
+    'An &lbrace;apple_fruit&rbrace;',
+  );
 
   expandHtml(
     'An &lbrace;apple_fruit&rbrace;',
@@ -68,13 +67,13 @@ test('expand2', function (t) {
     'An <a href="http://www.apple.com">Apple</a>',
   );
 
-  // TODO: Shouldn't perform link interpolation with expand2text.
-  // expandText(
-  //   'An {apple_fruit|Apple}',
-  //   {apple_fruit: 'http://www.apple.com'},
-  //   'An {apple_fruit|Apple}',
-  // );
-  // t.ok(/unexpected token/.test(error));
+  // Shouldn't perform link interpolation with expand2text.
+  expandText(
+    'An {apple_fruit|Apple}',
+    {apple_fruit: 'http://www.apple.com'},
+    'An {apple_fruit|Apple}',
+  );
+  t.ok(/unexpected token/.test(error));
 
   expandHtml(
     'An <a href="/apple">Apple</a>',
@@ -82,13 +81,13 @@ test('expand2', function (t) {
     'An <a href="/apple">Apple</a>',
   );
 
-  // TODO: HTML should be parsed as plain text with expand2text.
-  // expandText(
-  //   'An <a href="/apple">Apple</a>',
-  //   null,
-  //   'An <a href="/apple">Apple</a>',
-  // );
-  // t.equal(error, '');
+  // HTML should be parsed as plain text with expand2text.
+  expandText(
+    'An <a href="/apple">Apple</a>',
+    null,
+    'An <a href="/apple">Apple</a>',
+  );
+  t.equal(error, '');
 
   expandHtml(
     'A {apple_fruit|darn {apple}}',
@@ -151,7 +150,7 @@ test('expand2', function (t) {
   expandText('{x:%|}', {x: '&percnt;'}, '&percnt;');
   expandHtml('{x:%|}', {x: <p>hi</p>}, '<p>hi</p>');
   expandText('{x:a%c|}', {x: 'b'}, 'abc');
-  // TODO: expandText('{x:a&percnt;c|}', {x: 'b'}, 'a&percnt;c');
+  expandText('{x:a&percnt;c|}', {x: 'b'}, 'a&percnt;c');
   expandHtml('{x:a&percnt;c|}', {x: 'b'}, 'a%c');
 
   expandHtml('<a href="{x}"></a>', {x: '/&'}, '<a href="/&amp;"></a>');
